fix(policy): fix policy readiness probe delay
The policy container is configured differently than all other controllers: other
controllers configure an `initialDelaySeconds` on their `livenessProbe` but not
on their `readinessProbe`. The policy container, however, includes this
configuration on its `readinessProbe` but not on its `livenessProbe`.

This commit fixes the policy container to match the other controllers.

This substantially reduces the time it takes for the policy container to become
ready.