fix(policy): fix policy readiness probe delay (#13380)
The policy container is configured differently than all other controllers: other
controllers configure an `initialDelaySeconds` on their `livenessProbe` but not
on their `readinessProbe`. The policy container, however, includes this
configuration on its `readinessProbe` but not on its `livenessProbe`.

This commit fixes the policy container to match the other controllers.

This reduces pod readiness time from 20s to 4s.