Simplify cni config
This change removes the `policy` entry from the cni config template,
which isn't used. That contained a `__SERVICEACCOUNT_TOKEN__`
placeholder, which was coupling this config file with the
`ZZZ-linkerd-cni-kubeconfig` file generated by linkerd-cni. An upcoming
PR will add support for detecting changes in the mounted serviceaccount
token file (see #12573), and the current change will facilitate that
effort.