chore: simplify protoc scripts
The bin/protoc script is ancient and not useful, especially in light of tools
provided by dev containers. Furthermore, it includes a reference to an old gross
sourcefource downlaod page for unzip.

Furthermore, the protoc-go.sh script does seperate version pinning of build
dependencies.

To simplify all of this, this change removes the bin/protoc and bin/protoc-diff
scripts. The "proto" CI workflow is updated to use the protoc-go.sh script
directly.