chore: remove bin/helm-doc
The version in `bin/helm-doc` isn't in sync with what's installed in the
dev container, which generates a diff in the README chart files. CI uses
the dev container and it appears most folks do as well, so I'm removing
that script instead of having to keep it in sync.