test(policy): Reduce duplication in outbound policy API tests (#13543)
This change refactors the outbound policy API integration tests to reduce duplication and to be more principled about what combinations of routes and parents are tested.  The outbound policy API supports many route types (Linkerd HTTPRoute, gateway HTTPRoute, GRPCRoute, TCPRoute, and TLSRoute) and each of these routes may be attached to up to two parent types (Service and EgressNetwork).  Many behaviors of the outbound policy API are shared between these route and parent types and in order to test these behaviors, many tests were duplicated for each combination of types.  This resulted in a large amount of duplication and made it difficult to determine exactly which combinations were being tested and if any combinations were missed.  It also required that the duplicated tests be kept in sync and allowed the possibility that the tests could accidentally drift apart over time resulting in inconsistent test coverage.

We introduce the TestRoute and TestParent traits which capture the common properties of routes and parents respectively and refactor the outbound policy API tests to be generic over these types and then explicitly invoke the test with the combinations of types to which the tested behavior applies.  In cases where the tested behavior is specific to a particular type rather than being common to all types, a type specific test remains.

Signed-off-by: Alex Leong <alex@buoyant.io>