fix(policy): Update outbound policy routes even when parent ref has no port (#13584)
The `parent_ref` of an xRoute resource can optionally specify a port.  If no port is specified, the route applies to all ports.  These routes with `parent_ref` which do not specify a port are stored differently than routes which do, since these routes need to be applied for to all watches on any port on the parent.  When services are updated/created, we perform a reindex to ensure that all routes have the correct information about their backends.  However, routes which are stored in the separate way were not being updated in this case.  This means that routes with a `parent_ref` with no port can, depending on the sequence of operations, have stale information about their backend services.

We update these routes during the reindex operation to ensure they have up to date information about their backends and add a test for this behavior.

Signed-off-by: Alex Leong <alex@buoyant.io>