fix(cli): Don't require gateway CRDs in linkerd install
In https://github.com/linkerd/linkerd2/pull/12917 we made the Gateway API CRDs optional in Linkerd and updated the `linkerd check` command to no longer error if these CRDs were not present.  However, we missing making the corresponding change in `linkerd install`.  As a result, `linkerd install` returns an error if the Gateway API CRDs are not installed on the cluster.

We add the corresponding change to `linkerd install` so that it no longer returns an error if the Gateway API CRDs are not installed.

Signed-off-by: Alex Leong <alex@buoyant.io>