refactor(policy): handle opaque ports as an app protocol (#13659)
Currently, we store whether or not a port in a parent is opaque with a simple boolean. This works, but is somewhat restrictive if we want to specifiy different port protocols in the resource definition.

This converts it to a proper `AppProtocol` enum, which will allow us to use different protocol specifications for ports in the future.

Signed-off-by: Scott Fleener <scott@buoyant.io>