fix(policy): retry lease creation errors (#13677)
When the policy controller starts up, it attempts to create a Lease in the
cluster. If this fails with an unexpected error, the error is considered fatal
and the controller exits, causing the container to be restarted.

This can cause a substantial amount of CI flakiness (our CI checks for container
restarts), and it's generally undesirable for users to see container restarts.

To avoid this problem, this change adds simple retry functionality to the lease
creation.