test(policy): add E2E test for service `appProtocol` handling (#13737)
This adds some fairly straightforward tests for this functionality. It primarily verifies that a) given the presence of `appProtocol`, the policy for a service will always use only that protocol, and b) the proxy actually uses said protocol and doesn't attempt protocol detection. For instance, trying to use http/2 on an http/1 only workload will result in protocol errors.

Signed-off-by: Scott Fleener <scott@buoyant.io>