feat(crds): add installGatewayAPI helm chart value (#13739)
We add a new value to the `linkerd-crds` Helm chart called `installGatewayAPI` which acts as a default value for `enableHttpRoutes`, `enableTlsRoutes`, and `enableTcpRoutes`.

We also update the logic of the `linkerd install` and `linkerd upgrade` commands to set this `installGatewayAPI` value to true if there are any gateway API CRDs managed by Linkerd on the cluster and false otherwise.  Users can still override this setting by specifying the `--set installGatewayAPI=(false|true)` flag.

Signed-off-by: Alex Leong <alex@buoyant.io>