fix(helm): correct installGatewayAPI helm value defaulting (#13759)
The Helm function `default` will treat a boolean false value as unset and use the default instead of the false, even when it is set.  When rendering CRDs during install or upgrade, this can cause Linkerd to fall back to using the `installGatewayAPI` value even when `enableHttpRoutes` is explicitly set to false.

We replace the `default` function with a ternary which checks if the key is present.  We also add tests for both CLI and Helm.

Signed-off-by: Alex Leong <alex@buoyant.io>