feat!(multicluster): Federated services take metadata from member with the oldest Link (#13783)
When the service mirror controller detects the first member of a federated service, it will create the federated service itself and will copy the port definitions from the founding member service.  However, from that point on, the port definition in the federated service will never be updated, even if new members that define different ports join the federated service or if the ports in the founding member service are updated.

This leads to a scenario where the federated service’s port definitions become out of date if the member services’ port definitions change.  Similarly for labels and annotations.

We update the service mirror controller to look at the createdAt timestamp of each member service's Link resource and to use the member with the oldest Link as the authoritative source of truth for the federated service metadata: labels, annotations, and ports.

In the case where the service with the oldest Link leaves the federated service, it's metadata will continue to be present on the federated service until a resync occurs from the member with the next oldest Link (at most 10 minutes).

Signed-off-by: Alex Leong <alex@buoyant.io>