feat(multicluster): add link-gen command, deprecate link and unlink commands (#13858)
As of version 2.18 the golden path for establishing links will be by setting the `controllers` entry in the helm values, and applying the Link resource along with the credential secrets (see #13770).

This change deprecates the `linkerd multicluster link` command and puts in its place `linkerd multicluster link-gen` which only generates the Link resource and the credentials secrets. This is what the recently introduced flag `linkerd multicluster link --service-mirror=false` flag was accomplishing, but now we'd rather revert the introduction of that flag and instead deprecate the whole subcommand and introduce a new one.

Likewise, `linkerd mc unlink` is deprecated as well.