fix bug related to concurrent updates, resource versions, and event requeues (#13849)
When the service mirror controller encounters an error during processing an event, it requeues that event to be processed again up to a certain number of tries.  These events can store Service and Endpoint resources.

When two events which each cause updates to a mirror service happen in quick succession, one of these updates may fail because the other one has incremented the mirror service's resource version.  This manifests as a "Failed Precondition" during the update.  This update failure causes the event to be requeued; however the requeued event still holds the out-of-date service resource.  This means that when the requeued event is processed again, it will fail again because the service it holds is still out-of-date.  These failures will continue until the maximum number of requeue tries is reached and then we give up on the event and miss out on the update, leaving the mirror service in an incorrect state.

To fix this, we stop storing these resources in the event, and instead re-fetch them when processing the event.  This way, if an update fails due to a conflict and the event is requeued, then when the requeued event is processed, the mirror service will be fetched again and will be an up-to-date base for updates to be applied to.