feat(multicluster): add --only-controller flag to `linkerd mc unlink` (#13874)
* feat(multicluster): add --only-controller flag to `linkerd mc unlink`

This flag allows removing the service-mirror controller deployment and
all the associated resources for a given target cluster. It doesn't
output entries for the link CR, credential secrets, lease object or
mirror services. It can be used to migrate from controllers generated
by `linkerd mc link`, to the controllers managed by the multicluster
extension introduced in 2.18.