fix(test): avoid duplicate registry errors
The destinaation controller's cluster store registers a gague in its
constructor. When this constructor is called multiple times (i.e. in tests),
this can lead to a panic.

To avoid this panic, this change updates NewClusterStoreWithDecoder to accept a
prometheus registry). The NewClusterStore constructor (used by the application's
main) continues to use the default registry, but tests now construct their own
temporary registries to avoid duplicate registration errors.