test(CI): move go test timeout to justfile (#13927)
The go-test CI workflow specifies at 10m timeout.  However, if a test exceeds this timeout (for example if it is deadlocked) then CI terminates the job and the go test output doesn't give us any clues about which test specifically was running.  This makes it hard to identify what caused the failure.

We move the 10m timeout from the CI workflow into the `go-test` justfile recipe.  This causes `go test` to exit with a more informative error message once the timeout is reached, instead of being terminated by CI.

Signed-off-by: Alex Leong <alex@buoyant.io>

---------

Signed-off-by: Alex Leong <alex@buoyant.io>