fix(CLI): Check that the Gateway API version includes v1 resources (#13938)
When Linkerd is installed with an externally managed Gateway API, those externally managed CRDs must include the v1 resources versions of the HTTPRoute and GRPCRoute resources.  These resource versions are present in gateway API versions v1.0.0 and onward.

During install and upgrade, if using an externally managed Gateway API, we add a check to verify that the Gateway API includes v1 resource versions of HTTPRoute and GRPCRoute.  We return an error if these resource versions are not present.

```
> linkerd install --crds
the httproutes.gateway.networking.k8s.io CRD is missing the v1 version, please upgrade to Gateway API v1.0.0 or later
```

Signed-off-by: Alex Leong <alex@buoyant.io>