Set correct default port for tracing (#13937)
Now that the default tracing protocol is OpenTelemetry, this changes the default port for traces to the OpenTelemetry port on the collector instead of the OpenCensus one.

The current default port, matched with the current default trace protocol of OpenTelemetry, is currently broken as it causes traces to be sent to a collector port that expects OpenCensus traces. This is technically a breaking change, but it is less breaking than the change of the default protocol to OpenTelemetry.

More explicitly, if a user only used the defaults, this change brings them from a broken state to a working state. If a user brings their own tracing infrastructure with a custom collector address, this change doesn't affect them at all. The only users that may be broken by this are ones that explicitly set the protocol to OpenCensus, but we expect this to be rare as OpenCensus as a protocol has been sunset for a few years now.