fix(policy): include TCPRoute policy on opaque appProtocol services (#13948)
When a Service has a port marked with `appProtocol: linkerd.io/opaque`, the policy controller returns only the default opaque route and any TCPRoutes attached to that service and port are ignored.

We include any TCPRoute in the outbound policy when the `appProtocol` is marked as opaque.  

Furthermore, we update the logic to treat any unknown value for `appProtocol` as opaque (when previously, an unknown value was ignored, allowing any route type to be attached).  The consequence of this is TCPRoute is the only route type that can be attached to a pork with an unknown value for `appProtocol`.

Yet furthermore, we updated the logic to allow GRPCRoutes to be attached to ports marked with `appProtocol: kubernetes.io/h2c`.  This means that either GRPCRoutes or HTTPRoutes may be attached to such ports and the more specific type (GRPCRoute) will take higher precedence.

---------

Signed-off-by: Alex Leong <alex@buoyant.io>