build: validate edge version in release workflow (#13993)
Currently when an edge version is tagged, it gets validated only until
the chart_deploy job. If it fails, the release CLI gets published but
the chart not, leaving things in a mixed bad state.

This change validates the version early on in the release.yml workflow.

It repurposes the script bin/helm-bump-edge as bin/compute-edge-version
which performs the validation, and optionally mutates the version in the
helm charts (the latter being used only in the chart_deploy job).

The following tests the functionality, assuming the tip of the repo is tagged as
`edge-25.5.1` and a new tag is expected to be `edge-25.5.2`:

```bash
# Remove current tag
$ git tag -d edge-25.5.1

# Test with invalid tag
$ git tag edge-25.5.3
$ bin/compute-edge-version
Tag (edge-25.5.3) doesn't match computed edge version (edge-25.5.2)

# Test with valid tag
$ git tag -d edge-25.5.3
Deleted tag 'edge-25.5.3' (was 4823b7af3)
$ git tag edge-25.5.2
# Empty results means success
$ bin/compute-edge-version

# Mutate charts
$ bin/compute-edge-version update-charts
Bumping charts/linkerd2-cni/Chart.yaml to 2025.5.2
Bumping charts/linkerd-control-plane/Chart.yaml to 2025.5.2
Bumping charts/linkerd-crds/Chart.yaml to 2025.5.2
Bumping jaeger/charts/linkerd-jaeger/Chart.yaml to 2025.5.2
Bumping multicluster/charts/linkerd-multicluster/Chart.yaml to 2025.5.2
Bumping viz/charts/linkerd-viz/Chart.yaml to 2025.5.2
```