chore(CLI): Remove unused alpha CLI subcommand
The Linkerd CLI has a `alpha` subcommand that was previously used to hold experimental subcommands such as ones related to SMI.  All experimental subcommands have been removed and therefore the `alpha` subcommand is empty. This causes a pointless section to be printed at the end of the linkerd CLI usage help:

```
[...]
Additional help topics:
  linkerd alpha        experimental subcommands for Linkerd

Use "linkerd [command] --help" for more information about a command.
```

We remove this empty command so that it no longer shows up in the usage help:

```
Use "linkerd [command] --help" for more information about a command.
```

Signed-off-by: Alex Leong <alex@buoyant.io>