feat(helm): grant linkerd-config read access to additional SAs (#14092)
The intent of this change is to enable extensions to access Linkerd's configuration.

It introduces the linkerd-config RoleBinding, associated to the existing Role `ext-namespace-metadata-linkerd-config` that grants read access to the linkerd-config ConfigMap (that Role name is now outdated but leaving as-is to avoid disruption).

The service accounts for that binding are to be listed on the new `values.yaml` entry `configReaders`, for example:

```yaml
configReaders:
- name: ext1
  namespace: linkerd-ext1
- name: ext2
  namespace: linkerd-ext2
```