chore(build): remove choco (#14234)
The chocolatey package was being built only for releases with a `stable`
tag, which is no longer produced in this repo. And CI started warning
about the usage of the deprecated windows-2019 runner it relies on.

If there's need for this type of package, it should probably be
maintained outside this repo, like it's done with the brew linkerd
formula for example.