fix(helm): Fix helm rendering error when upgrading #14244
Fixes #14228 

When upgrading from a previous version of Linkerd, the `proxy.tracing` key will not exist in the Linkerd config ConfigMap.  If a user attempts to use the inject CLI command in `--manual` mode, the config will be loaded from the Linkerd config ConfigMap and then rendering of the pod patch template will fail:

```
template: patch/templates/patch.json:106:16: executing "patch/templates/patch.json" at <.Values.proxy.tracing.enable>: nil pointer evaluating interface {}.enable
```

This is because the template is attempting to access the `enabled` key in the tracing struct which is nil.

We add guards using the `dig` Helm template function so that template rendering does not fail, even if the `proxy.tracing` key is not present.

Signed-off-by: Alex Leong <alex@buoyant.io>