fix(helm): Fix cli version string in extensions (#14309)
Linkerd extensions have, in their deployment templates, a label like:

```
app.kubernetes.io/version: {{default .Values.linkerdVersion .Values.cliVersion}}
```

This can be a problem if the `cliVersion` value has been set because this value typically includes a space e.g. `"linkerd/cli edge-25.6.7"` and spaces are not legal in label values.  However, this has not been a problem because extensions have not been setting this value.  But because this value has not been set, the `created-by` annotation has not been correct when using the CLI:  

```
linkerd.io/created-by: linkerd/helm edge-25.6.7
```

We update this to set the `app.kubernetes.io/version` label to `.Values.linkerdVersion` in all cases which must not include a space.  We also set `.Values.cliVersion` for the extensions which correctly updates the created-by annotation to indicate that the manifest was rendered with the CLI rather than with Helm.

Signed-off-by: Alex Leong <alex@buoyant.io>