chore(inject): remove PatchProducer abstraction (#14346)
In a recent change, we introduced the `PatchProducer` abstraction in order to provide a way to add customizable injection logic to the injector. After some further experimentation, it has became clear that there might be better ways to customize the injection logic. 

For now, we remove this abstraction as we are not going to be using it in the near future.

Signed-off-by: Zahari Dichev <zaharidichev@gmail.com>