fix(ci): remove `windows_static_cli_tests`
this piece of our ci system runs the cli integration tests on windows,
but does so within the course of our release workflow rather than
regular ci.

this currently does not work because it cannot find its artifact.

this commit removes it from the release workflow. if we determine that
we would like to restore this, we should do so by placing this in the
ci action for pull requests to catch issues with this earlier.

Signed-off-by: katelyn martin <kate@buoyant.io>