fix(destination): Replace With with GetMetricWith (#14428)
There are a number of places in the destination controller where prometheus metrics are constructed using the `.With` function to supply label names.  The `.With` function panics if it encounters an invalid label name.  

To avoid potential panics, we replace all uses of `.With` with the safer `.GetMetricWith` which instead returns an error if it is provided an invalid label name.  We then handle these errors appropriately instead of panicking.

Signed-off-by: Alex Leong <alex@buoyant.io>