feat(gatewayAPI): Stop installing the gateway API by default (#14523)
We switch the default value of `installGatewayAPI` from true to false.

This has the effect that for Helm installs and upgrades, we will no longer install the gateway API CRDs by default and will instead depend on them already being installed on the cluster.

For Helm upgrades from 2.18, this is a graceful upgrade because 2.18 added the `helm.sh/resource-policy=keep` annotation to these CRDs.  This means that even though this default is changing, upgrading from 2.18 will not cause Helm to delete these CRDs.

This change has no effect on the Linkerd CLI which was already not installing gateway API CRDs by default.

Signed-off-by: Alex Leong <alex@buoyant.io>