chore(tests): Separate tracing tests from viz (#14547)
Now that the proxy has tracing configuration added to the proxy injector, we can decouple the tracing tests from the viz and jaeger extensions.

This creates a new set of top-level tests for tracing and moves the existing one from the viz integration tests, with a few modifications now that it doesn't have to install the extensions.

Signed-off-by: Scott Fleener <scott@buoyant.io>