feat!(tracing): Improve control plane tracing configuration
The current control plane tracing relies on the linkerd-jaeger extension, and does not work when using the native tracing configuration.

This removes the previous configs, and adds new control plane tracing config that mirrors the existing proxy tracing configs.

The previous configuration was meant entirely for internal testing purposes, and shouldn't be subject to any breaking change guarantees.

Signed-off-by: Scott Fleener <scott@buoyant.io>