feat(tracing)!: Improve control plane tracing configuration (#14557)
* chore(docs): Update tracing documentation

This clarifies the tracing documentation that the trace collector must be meshed, along with the specifics of how the service account name for the collector should be set.

Signed-off-by: Scott Fleener <scott@buoyant.io>

* chore(docs): Include `serviceAccountNamespace` for tracing collector

This separates out the service account namespace from the name to make it clearer how to set the correct service account for the trace collector.

Signed-off-by: Scott Fleener <scott@buoyant.io>

* chore(docs): Make docs clearer about what needs to be set for the collector mesh identity

Signed-off-by: Scott Fleener <scott@buoyant.io>

* feat!(tracing): Improve control plane tracing configuration

The current control plane tracing relies on the linkerd-jaeger extension, and does not work when using the native tracing configuration.

This removes the previous configs, and adds new control plane tracing config that mirrors the existing proxy tracing configs.

The previous configuration was meant entirely for internal testing purposes, and shouldn't be subject to any breaking change guarantees.

Signed-off-by: Scott Fleener <scott@buoyant.io>

* fix(test): Update tracing test values

Signed-off-by: Scott Fleener <scott@buoyant.io>

* feat(charts): Switch ServiceAccountName to separate Name and Namespace fields

Signed-off-by: Scott Fleener <scott@buoyant.io>

* fix: Change deprecated casing for lint

Signed-off-by: Scott Fleener <scott@buoyant.io>

* chore: Rename fields and semantics of tracing identity

Signed-off-by: Scott Fleener <scott@buoyant.io>

* chore: Fix bad auto-rename

Signed-off-by: Scott Fleener <scott@buoyant.io>

* chore: Change to breaking API change

Signed-off-by: Scott Fleener <scott@buoyant.io>

* chore: Regenerate goldens

Signed-off-by: Scott Fleener <scott@buoyant.io>

* fix: Use correct trace enable value

Signed-off-by: Scott Fleener <scott@buoyant.io>

* fix(docs): Typos in docs

Signed-off-by: Scott Fleener <scott@buoyant.io>

---------

Signed-off-by: Scott Fleener <scott@buoyant.io>