feat(inject): Populate proxy tracing values from annotations
This adds functionality to the proxy injector that parses tracing semantic convention annotations and labels, and populates those into the trace labels provided to the proxy.

This also updates the set of extra labels (pod IP, pod ID, etc.) to be part of the control plane values, which both allows user customization and the proxy injector to modify those values directly.

Signed-off-by: Scott Fleener <scott@buoyant.io>