feat(inject): Populate proxy tracing values from annotations (#14592)
* feat: Add label for root parent object

Signed-off-by: Scott Fleener <scott@buoyant.io>

* fix: Use correct format for API resource

The GroupVersionResource requires the resource to be both lowercase and plural.

Signed-off-by: Scott Fleener <scott@buoyant.io>

* chore(docs): Add godoc for RootOwnerRetrieverFunc

Signed-off-by: Scott Fleener <scott@buoyant.io>

* chore(inject): Add group label for root parent

Signed-off-by: Scott Fleener <scott@buoyant.io>

* feat(inject): Populate proxy tracing values from annotations

This adds functionality to the proxy injector that parses tracing semantic convention annotations and labels, and populates those into the trace labels provided to the proxy.

This also updates the set of extra labels (pod IP, pod ID, etc.) to be part of the control plane values, which both allows user customization and the proxy injector to modify those values directly.

Signed-off-by: Scott Fleener <scott@buoyant.io>

* test(inject): Update webhook tests for root owner annotations

Signed-off-by: Scott Fleener <scott@buoyant.io>

* chore(inject): Fix test failures

Signed-off-by: Scott Fleener <scott@buoyant.io>

---------

Signed-off-by: Scott Fleener <scott@buoyant.io>