fix(destination): avoid blocking the endpoint translators
The synchronized get stream, used by the federated service watcher, uses an
unbuffered channel. This means that if the stream.Send blocks, we will block the
endpoint translator as well as any attempts to Stop() the watcher! This is an
obvious avenue for leaking goroutines.

This change updates these channels to be buffered and updates Send and Stop to
gracefully handle the overflow cases.