feat(destionation): add --stream-queue-capacity arg
The destination controller has a hardcoded stream queue capacity of 100. This
commit introduces a new command-line argument `--stream-queue-capacity` that
allows users to customize this value according to their needs.

We believe that '1' is a better default value for most use cases, as it reduces
memory consumption and latency. However, we want to be incremental about this
change. Therefore, this configuration allows for us to test this new default
value in a controlled manner.